refactor(backend): modularize oversized config module into app/config package
Decoupled the monolithic app/config.py to reduce cognitive load:
- app/config/settings.py: Contains application settings and environment variables
- app/config/models.py: Contains all Pydantic configuration models
- app/config/loader.py: Handles reading YAML configuration and managing global state
- app/config/__init__.py: Exposes everything to preserve backwards compatibility

Verified with passing unit tests.
